separated out /tests into its own directory
